refactor: split dispatch/wasm.rs helpers into focused submodules
Behavior-preserving extraction from the 2,343-line WASM dispatch module.
The entangled core dispatch state machine stays in wasm.rs (1,177 lines);
the orthogonal helpers move out verbatim:

  wasm/phases.rs            121  WASM_DISPATCH_PHASE_* constants + phase
                                 span/timing instrumentation helpers
  wasm/interceptors.rs      183  local blob / file-value HTTP interceptors
  wasm/llm_observability.rs 943  LLM wide-events, gen_ai span attrs,
                                 LLMObs span submission, error recording

Only visibility widened (private -> pub(super); one
pub(in crate::state::dispatch) + re-export keeps invocation_artifacts.rs
compiling unchanged). 9 tests relocated with their subjects; all
pre-existing same-line determinism-ok annotations preserved. The
observability contract test now includes the phases submodule source,
since the phase-name constants live there.

385 temper-server lib tests green; determinism hook passes on all files.
DST review: PASS. Code review: PASS (verbatim-move verified per function).
